# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: JAAS Module 0 Identity API
  description: API specifications for Module 0 — Tenant, Identity, and RBAC authorization core.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Dev Server
paths:
...
  /roles:
    post:
      summary: Create Custom Role
      tags:
        - Roles
      security:
        - BearerAuth: []
        - TenantSubdomain: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                permission_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '201':
          description: Role created
    get:
      summary: List Tenant Roles
      tags:
        - Roles
      security:
        - BearerAuth: []
        - TenantSubdomain: []
      responses:
        '200':
          description: List of roles
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your user access token.
    TenantSubdomain:
      type: apiKey
      in: header
      name: Host
      description: The request host header containing the tenant subdomain slug (e.g. acme.jaas.com).